common: sec_proxy: Add per device secure proxy base_address
There are two communication scheme that have been enabled to communicate
with Secure Proxy in TI.
a) A full fledged prioritized communication scheme, which involves upto
   5 threads from the perspective of the host software
b) A much simpler "lite" version which is just a two thread scheme
   involving just a transmit and receive thread scheme.

The (a) scheme is used in AM654, J721e, J7200 platforms where as
(b) scheme is used in AM64 platforms. Based on this, the base address
for secure proxy are different in platforms using scheme (a) vs scheme (b).
Define generic and lite secure proxy base addresses and attach to the
respective SoCs.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>