soc: remove array indexing from soc data
When new features are added to the tisci they typically will be placed
inside the reserved spaces in the abi. This has the side effect of
reordering large portions of our data in k3conf when everything below
the additions are shifted to make room.

To make updates to the tisci easier to diff in the future, remove the
array indexing from all tisci soc data in k3conf.

There are no functional changes with this patch.

This patch was the output of:

  $ find soc/ -iname '*.c' -exec sed -i 's/\[[0-9]\+\] = //g' {} \;

Signed-off-by: Bryan Brattlof <bb@ti.com>