cmd: read/write: Fix the kernel crash issue on reading 32bit registers
k3conf read command leads to kernel crash on reading few 32bit registers.
This is because it always reads 64 bits at the specified register address
irrespective of the register width. For example, on j721e "k3conf read
0x43000050" leads to kernel crash because the k3conf tool tries to read 64
bits but the register width is 32 bits and the address of the next valid
register is at 0x43000060 due to which 0x43000054 - 0x43000059 addresses
are invalid which leads to kernel crash.

Inorder to fix the issue an optional argument named size is added to both
the k3conf read and write commands. The user can specify the number of
bits to be read or written in the size argument. For example,
"k3conf read 0x43000050 32", "k3conf write 0x43000050 0x10 32".

Signed-off-by: Thejasvi Konduru <t-konduru@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>