common: socinfo: Do not accept REV_PG_MAX as valid revision
This revisions are zero-indexed, and REV_PG_MAX is the entry one
after the last valid entry. Do not accept REV_PG_MAX as valid.
This means we also do not need to define that value in the table.

While here remove hard-coded limits on this string size, let
the compiler size them based on the strings themselves in memory.

Signed-off-by: Andrew Davis <afd@ti.com>