common: socinfo: Remove soc_full_name and strncat usage
Usage of strncat here causes warnings due to the possibility of not
having space for the null terminator when the string is the full
length of the destination buffer.

Simply remove the use of strncat and the need for a sized buffer
by directly assigning the const strings to a pointer and using
that in the table.

Signed-off-by: Andrew Davis <afd@ti.com>