common: socinfo: Remove soc and rev from k3conf_soc_info
This are not used outside of socinfo.c and can be made local.

We also move soc_revision_j721e and soc_revision_generic for the same
reasons.

Functions soc_is_j721e() and soc_is_am654() are not used and by removing
them we can make soc and rev local to the soc_init() function.

Signed-off-by: Andrew Davis <afd@ti.com>