common: ddr-perf: Fix count overflow off-by-one
Currently complex logic is being used to try to deal with the overflow
case when the 32bit counters rollover. Not only is this logic wrong
as it causes an off-by-one error on rollover, it isn't needed at all.
Unsigned integer overflow/underflow is defined behavior and correctly
handles the rollover case with simple subtraction. Do this here.

Signed-off-by: Andrew Davis <afd@ti.com>